# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/valorant/account/{puuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает список скинов игрока с именем и иконкой.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Valorant"
                ],
                "summary": "Получить скины Valorant аккаунта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игрока",
                        "name": "puuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Принудительное обновление",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/valorant.getAccountResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/valorant.getAccountResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/valorant.getAccountResponse"
                        }
                    }
                }
            }
        },
        "/v1/valorant/matches/{puuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает последние матчи со статистикой игроков и дуэлями.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Valorant"
                ],
                "summary": "История матчей игрока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игрока",
                        "name": "puuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Принудительное обновление (игнорировать кэш)",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/valorant.getMatchHistoryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/valorant.getMatchHistoryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/valorant.getMatchHistoryResponse"
                        }
                    }
                }
            }
        },
        "/v1/valorant/mmr/{puuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает информацию о ранге, рейтинге и последнем конкурентном обновлении игрока",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Valorant"
                ],
                "summary": "Получить MMR и ранг игрока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игрока",
                        "name": "puuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "MMR успешно получен",
                        "schema": {
                            "$ref": "#/definitions/valorant.getMMRResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован - сессия истекла",
                        "schema": {
                            "$ref": "#/definitions/valorant.getMMRResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/valorant.getMMRResponse"
                        }
                    }
                }
            }
        },
        "/v1/valorant/store/{puuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Получает информацию о ежедневном магазине для конкретного пользователя по его PUUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Valorant"
                ],
                "summary": "Получить ежедневный магазин пользователя",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Принудительное обновление (игнорировать кэш)",
                        "name": "force",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "UUID игрока",
                        "name": "puuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Магазин успешно получен",
                        "schema": {
                            "$ref": "#/definitions/valorant.getDailyStoreResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован - сессия истекла или недействительна",
                        "schema": {
                            "$ref": "#/definitions/valorant.getDailyStoreResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/valorant.getDailyStoreResponse"
                        }
                    }
                }
            }
        },
        "/v1/valorant/wallet/{puuid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает количество Valorant Points (VP), Radianite Points и Kingdom Credits игрока",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Valorant"
                ],
                "summary": "Получить баланс кошелька игрока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игрока",
                        "name": "puuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Баланс успешно получен",
                        "schema": {
                            "$ref": "#/definitions/valorant.getWalletResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован - сессия истекла",
                        "schema": {
                            "$ref": "#/definitions/valorant.getWalletResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/valorant.getWalletResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "message": {
                    "type": "string"
                },
                "meta": {}
            }
        },
...
        "valorant.bundleDTO": {
            "type": "object",
            "properties": {
                "expires_in_seconds": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.bundleItem"
                    }
                },
                "name": {
                    "type": "string"
                },
                "total_base_price": {
                    "type": "integer"
                },
                "total_discount_percent": {
                    "type": "number"
                },
                "total_discounted_price": {
                    "type": "integer"
                }
            }
        },
        "valorant.bundleItem": {
            "type": "object",
            "properties": {
                "base_price": {
                    "type": "integer"
                },
                "discount_percent": {
                    "type": "number"
                },
                "discounted_price": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "is_promo": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "valorant.getAccountResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "skins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.skinDTO"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "valorant.getDailyStoreResponse": {
            "type": "object",
            "properties": {
                "accessories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.storeItem"
                    }
                },
                "bundles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.bundleDTO"
                    }
                },
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "store": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.storeItem"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "valorant.getMMRResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "mmr": {
                    "$ref": "#/definitions/valorant.mmrInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "valorant.getMatchHistoryResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.matchDTO"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "valorant.getWalletResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "wallet": {
                    "$ref": "#/definitions/wallet.Wallet"
                }
            }
        },
        "valorant.matchDTO": {
            "type": "object",
            "properties": {
                "duration_ms": {
                    "type": "integer"
                },
                "kills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.matchKillDTO"
                    }
                },
                "map_id": {
                    "type": "string"
                },
                "match_id": {
                    "type": "string"
                },
                "played_at": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/valorant.matchPlayerDTO"
                    }
                },
                "queue_id": {
                    "type": "string"
                },
                "team_blue_rounds_won": {
                    "type": "integer"
                },
                "team_red_rounds_won": {
                    "type": "integer"
                },
                "team_red_won": {
                    "type": "boolean"
                }
            }
        },
        "valorant.matchKillDTO": {
            "type": "object",
            "properties": {
                "assistants": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "killer": {
                    "type": "string"
                },
                "round": {
                    "type": "integer"
                },
                "victim": {
                    "type": "string"
                }
            }
        },
        "valorant.matchPlayerDTO": {
            "type": "object",
            "properties": {
                "agent_icon": {
                    "type": "string"
                },
                "agent_name": {
                    "type": "string"
                },
                "assists": {
                    "type": "integer"
                },
                "character_id": {
                    "type": "string"
                },
                "deaths": {
                    "type": "integer"
                },
                "kills": {
                    "type": "integer"
                },
                "name": {
                    "description": "\"GameName#TagLine\"",
                    "type": "string"
                },
                "puuid": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "team": {
                    "type": "string"
                }
            }
        },
        "valorant.mmrInfo": {
            "type": "object",
            "properties": {
                "last_map_id": {
                    "type": "string"
                },
                "last_match_id": {
                    "type": "string"
                },
                "ranked_rating": {
                    "type": "integer"
                },
                "rr_change": {
                    "type": "integer"
                },
                "tier": {
                    "type": "integer"
                }
            }
        },
        "valorant.skinDTO": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "valorant.storeItem": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "wallet.Wallet": {
            "type": "object",
            "properties": {
                "kingdom_credits": {
                    "type": "integer"
                },
                "radianite_points": {
                    "type": "integer"
                },
                "valorant_points": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}